+++ app/src/main/java/com/example/fse_project/domain/model/Vehicle.kt
@@ -3,8 +3,8 @@ package com.example.fse_project.domain.model
 import com.example.fse_project.data.local.database.entity.ConnectorType
 
 data class Vehicle(
-    val id : Int,
-    val ownerId : Int, //user
+    val id : Long,
+    val ownerId : Long, //user
     val brand : String,
     val model : String,
     val capacity : Int,
